REF: Move pub/sub event handler into `octue.cloud.pub_sub.events`
