+++ octue/cloud/pub_sub/service.py
@@ -16,8 +16,7 @@ from google.cloud import pubsub_v1
 import octue.exceptions
 from octue.cloud.events.validation import raise_if_event_is_invalid
 from octue.cloud.pub_sub import Subscription, Topic
-from octue.cloud.pub_sub.event_handler import GoogleCloudPubSubEventHandler
-from octue.cloud.pub_sub.events import extract_event_and_attributes_from_pub_sub_message
+from octue.cloud.pub_sub.events import GoogleCloudPubSubEventHandler, extract_event_and_attributes_from_pub_sub_message
 from octue.cloud.pub_sub.logging import GoogleCloudPubSubHandler
 from octue.cloud.service_id import (
     convert_service_id_to_pub_sub_form,
