+++ tests/cloud/pub_sub/test_events.py
@@ -12,7 +12,7 @@ from octue.cloud.emulators._pub_sub import (
     MockTopic,
 )
 from octue.cloud.emulators.child import ServicePatcher
-from octue.cloud.pub_sub.event_handler import GoogleCloudPubSubEventHandler
+from octue.cloud.pub_sub.events import GoogleCloudPubSubEventHandler
 from octue.resources.service_backends import GCPPubSubBackend
 from tests import TEST_PROJECT_NAME
 from tests.base import BaseTestCase
@@ -39,7 +39,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """Test that a TimeoutError is raised if message handling takes longer than the given timeout."""
         question_uuid, _, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(
                 subscription=mock_subscription,
                 receiving_service=parent,
@@ -54,7 +54,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """Test that messages received in order are handled in order."""
         question_uuid, mock_topic, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(
                 subscription=mock_subscription,
                 receiving_service=parent,
@@ -86,7 +86,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """Test that messages received out of order are handled in order."""
         question_uuid, mock_topic, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(
                 subscription=mock_subscription,
                 receiving_service=parent,
@@ -139,7 +139,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """
         question_uuid, mock_topic, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(
                 subscription=mock_subscription,
                 receiving_service=parent,
@@ -190,7 +190,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """Test that message handling works with no timeout."""
         question_uuid, mock_topic, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(
                 subscription=mock_subscription,
                 receiving_service=parent,
@@ -230,7 +230,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """Test that a delivery acknowledgement message is handled correctly."""
         question_uuid, mock_topic, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(subscription=mock_subscription, receiving_service=parent)
 
         child = MockService(backend=GCPPubSubBackend(project_name=TEST_PROJECT_NAME))
@@ -256,7 +256,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """Test that an error is raised if a heartbeat isn't received before a heartbeat is first checked for."""
         question_uuid, _, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(subscription=mock_subscription, receiving_service=parent)
 
         with self.assertRaises(TimeoutError) as error:
@@ -269,7 +269,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """Test that an error is raised if heartbeats stop being received within the maximum interval."""
         question_uuid, _, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(subscription=mock_subscription, receiving_service=parent)
 
         event_handler._last_heartbeat = datetime.datetime.now() - datetime.timedelta(seconds=30)
@@ -283,7 +283,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """Test that an error is not raised if a heartbeat has been received in the maximum allowed interval."""
         question_uuid, mock_topic, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(subscription=mock_subscription, receiving_service=parent)
 
         event_handler._last_heartbeat = datetime.datetime.now()
@@ -308,7 +308,7 @@ class TestPubSubEventHandler(BaseTestCase):
             child._send_message(message=message["event"], attributes=message["attributes"], topic=mock_topic)
 
         with patch(
-            "octue.cloud.pub_sub.event_handler.GoogleCloudPubSubEventHandler._time_since_last_heartbeat",
+            "octue.cloud.pub_sub.events.GoogleCloudPubSubEventHandler._time_since_last_heartbeat",
             datetime.timedelta(seconds=0),
         ):
             event_handler.handle_events(maximum_heartbeat_interval=0)
@@ -317,7 +317,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """Test that the time since the last heartbeat is `None` if no heartbeat has been received yet."""
         question_uuid, _, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(subscription=mock_subscription, receiving_service=parent)
 
         self.assertIsNone(event_handler._time_since_last_heartbeat)
@@ -342,7 +342,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """
         question_uuid, mock_topic, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(
                 subscription=mock_subscription,
                 receiving_service=parent,
@@ -395,7 +395,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """Test that missing messages in the middle of the event stream can be skipped."""
         question_uuid, mock_topic, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(
                 subscription=mock_subscription,
                 receiving_service=parent,
@@ -452,7 +452,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """Test that multiple blocks of missing messages in the middle of the event stream can be skipped."""
         question_uuid, mock_topic, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(
                 subscription=mock_subscription,
                 receiving_service=parent,
@@ -539,7 +539,7 @@ class TestPubSubEventHandler(BaseTestCase):
         """Test that the result message is still handled if all other messages are missing."""
         question_uuid, mock_topic, mock_subscription = create_mock_topic_and_subscription()
 
-        with patch("octue.cloud.pub_sub.event_handler.SubscriberClient", MockSubscriber):
+        with patch("octue.cloud.pub_sub.events.SubscriberClient", MockSubscriber):
             event_handler = GoogleCloudPubSubEventHandler(
                 subscription=mock_subscription,
                 receiving_service=parent,
